Commented out printing line in FastRTParser

+++ src/navigation/shared/Person.java
@@ -16,7 +16,7 @@ public class Person {
     private int SPF;
     private double bodyExposure; //as a percentage, 0-100, 0 = fully covered, 100 = no coverage
     
-    public Person(int skinType, int SPF, int bodyExposure) {
+    public Person(int skinType, int SPF, double bodyExposure) {
         this.skinType = skinType;
         this.SPF = SPF;
         this.bodyExposure = bodyExposure;
